# Conflicts:
#	sheep-common/pom.xml

+++ sheep-portal/src/main/java/sheep/portal/pojo/PaperList.java
@@ -22,7 +22,7 @@ public class PaperList<T> {
     /**
      * 总页码
      */
-    private Integer totalPages;
+    private Long totalPages;
 
 
 }
